% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_code_of_conduct.R
\name{add_code_of_conduct}
\alias{add_code_of_conduct}
\title{Add a CODE OF CONDUCT file}
\usage{
add_code_of_conduct(
  email = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{email}{a \code{character} of length 1. The email address of the user
(considered as the maintainer and code owner of the project).}

\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a \code{CODE_OF_CONDUCT.md} file adapted from the
Contributor Covenant, version 2.1 available at
\url{https://www.contributor-covenant.org/version/2/1/code_of_conduct.html}.
}
\examples{
\dontrun{
add_code_of_conduct()
}
}
\seealso{
Other create files:
\code{\link[=add_citation]{add_citation()}},
\code{\link[=add_codeowners]{add_codeowners()}},
\code{\link[=add_compendium]{add_compendium()}},
\code{\link[=add_contributing]{add_contributing()}},
\code{\link[=add_description]{add_description()}},
\code{\link[=add_dockerfile]{add_dockerfile()}},
\code{\link[=add_license]{add_license()}},
\code{\link[=add_makefile]{add_makefile()}},
\code{\link[=add_package_doc]{add_package_doc()}},
\code{\link[=add_readme_rmd]{add_readme_rmd()}},
\code{\link[=add_renv]{add_renv()}},
\code{\link[=add_testthat]{add_testthat()}},
\code{\link[=add_vignette]{add_vignette()}}
}
\concept{create files}
