% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contributing.R
\name{add_contributing}
\alias{add_contributing}
\title{Add a CONTRIBUTING file}
\usage{
add_contributing(
  email = NULL,
  organisation = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{email}{a \code{character} of length 1. The email address of the user
(considered as the maintainer and code owner of the project).}

\item{organisation}{A character of length 1. The name of the GitHub
organisation to host the package. If \code{NULL} (default) the GitHub account
will be used. This argument is used to set the URL of the package
(hosted on GitHub).}

\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a \code{CONTRIBUTING.md} file providing general guidelines
outlining the best way to contribute to the project (need to be adapted).
}
\examples{
\dontrun{
add_contributing()
}
}
\seealso{
Other create files:
\code{\link[=add_citation]{add_citation()}},
\code{\link[=add_code_of_conduct]{add_code_of_conduct()}},
\code{\link[=add_codeowners]{add_codeowners()}},
\code{\link[=add_compendium]{add_compendium()}},
\code{\link[=add_description]{add_description()}},
\code{\link[=add_dockerfile]{add_dockerfile()}},
\code{\link[=add_license]{add_license()}},
\code{\link[=add_makefile]{add_makefile()}},
\code{\link[=add_package_doc]{add_package_doc()}},
\code{\link[=add_readme_rmd]{add_readme_rmd()}},
\code{\link[=add_renv]{add_renv()}},
\code{\link[=add_testthat]{add_testthat()}},
\code{\link[=add_vignette]{add_vignette()}}
}
\concept{create files}
