% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependabot.R
\name{add_dependabot}
\alias{add_dependabot}
\title{Add a DEPENDABOT file}
\usage{
add_dependabot(overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a \code{dependabot.yaml} file (configuration file) in the
\verb{.github/} directory. This GitHub Action will be triggered once a week to
check for dependency updates (used by any GitHub Action of the repository).
If an update is available, this bot will open a Pull Request and user will
be invited to review changes.

This function is called by \code{\link[=add_github_action]{add_github_action()}} to automatically create
this configuration file if any GitHub Action is used.
}
\examples{
\dontrun{
add_dependabot()
}
}
\seealso{
Other development functions:
\code{\link[=add_dependencies]{add_dependencies()}},
\code{\link[=add_github_action]{add_github_action()}},
\code{\link[=add_issue_template]{add_issue_template()}},
\code{\link[=add_r_depend]{add_r_depend()}},
\code{\link[=add_to_buildignore]{add_to_buildignore()}},
\code{\link[=add_to_gitignore]{add_to_gitignore()}}
}
\concept{development functions}
