% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_badge.R
\name{add_github_actions_badge}
\alias{add_github_actions_badge}
\title{Add a R-CMD-CHECK badge}
\usage{
add_github_actions_badge(github = NULL, organisation = NULL, quiet = FALSE)
}
\arguments{
\item{github}{user GITHUB account where the project is hosted.}

\item{organisation}{(alternative) name of the GITHUB organisation where the
project is hosted.  If \code{NULL} (default) the GITHUB account will be used.
Otherwise the argument will be used.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function adds a R-CMD-CHECK badge to the \emph{README.Rmd}. This function
must be run after \code{\link[=add_github_actions_check]{add_github_actions_check()}} which will setup GitHub
Actions (checks and tests).
}
\examples{
\dontrun{
add_github_actions_badge()
}
}
\seealso{
Other adding badges: 
\code{\link{add_cran_badge}()},
\code{\link{add_dependencies_badge}()},
\code{\link{add_license_badge}()},
\code{\link{add_lifecycle_badge}()},
\code{\link{add_repostatus_badge}()}
}
\concept{adding badges}
