% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_check.R
\name{add_github_actions_check}
\alias{add_github_actions_check}
\title{Setup GitHub Actions to check and test package}
\usage{
add_github_actions_check(open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{open}{a logical value. If \code{TRUE} the file is opened in the editor.
Default is \code{FALSE}.}

\item{overwrite}{a logical value. If this file already exists and
\code{overwrite = TRUE}, its content will be erased.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function creates a configuration file (\code{.yaml}) to setup GitHub Actions
to check and test the package. This file will be added (from a template)
in the folder \code{.github/workflows/R-CMD-check.yaml}.
}
\details{
This workflow runs \verb{R CMD check} on the three major operating systems
(Ubuntu, macOS, and Windows) on the latest release of R.
}
\examples{
\dontrun{
add_github_actions_check()
}
}
\seealso{
Other development functions: 
\code{\link{add_dependencies}()},
\code{\link{add_github_actions_pkgdown}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
