% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_codecov.R
\name{add_github_actions_codecov}
\alias{add_github_actions_codecov}
\title{Setup GitHub Actions to report code coverage}
\usage{
add_github_actions_codecov(open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{open}{a logical value. If \code{TRUE} the file is opened in the editor.
Default is \code{FALSE}.}

\item{overwrite}{a logical value. If this file already exists and
\code{overwrite = TRUE}, its content will be erased.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function creates a configuration file (\code{.yaml}) to setup GitHub Actions
to report code coverage when testing the package. This file will be added
(from a template) in the folder \code{.github/workflows/test-coverage.yaml}.
}
\examples{
\dontrun{
add_github_actions_codecov()
}
}
\seealso{
Other development functions: 
\code{\link{add_dependencies}()},
\code{\link{add_github_actions_check}()},
\code{\link{add_github_actions_pkgdown}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
