% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_codecov_badge.R
\name{add_github_actions_codecov_badge}
\alias{add_github_actions_codecov_badge}
\title{Add a Test coverage badge}
\usage{
add_github_actions_codecov_badge(organisation = NULL, quiet = FALSE)
}
\arguments{
\item{organisation}{A character of length 1. The name of the GitHub
organisation to host the package. If \code{NULL} (default) the GitHub account
will be used.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
A badge as a markdown expression.
}
\description{
This function adds a \strong{Test coverage} badge to the \code{README.Rmd}. This
function must be run after \code{\link[=add_github_action]{add_github_action()}} which will setup
GitHub Actions to report the percentage of code cover by units tests.

Make sure that 1) a \code{README.Rmd} file exists at the project root and 2) it
contains a block starting with the line \verb{<!-- badges: start -->} and ending
with the line \verb{<!-- badges: end -->}.

Don't forget to re-render the \code{README.md}.
}
\examples{
\dontrun{
add_github_actions_codecov_badge()
}
}
\seealso{
Other adding badges:
\code{\link[=add_codecov_badge]{add_codecov_badge()}},
\code{\link[=add_cran_downloads_badge]{add_cran_downloads_badge()}},
\code{\link[=add_cran_total_download_badge]{add_cran_total_download_badge()}},
\code{\link[=add_cran_version_badge]{add_cran_version_badge()}},
\code{\link[=add_dependencies_badge]{add_dependencies_badge()}},
\code{\link[=add_github_actions_check_badge]{add_github_actions_check_badge()}},
\code{\link[=add_github_actions_pkgdown_badge]{add_github_actions_pkgdown_badge()}},
\code{\link[=add_license_badge]{add_license_badge()}},
\code{\link[=add_lifecycle_badge]{add_lifecycle_badge()}},
\code{\link[=add_repostatus_badge]{add_repostatus_badge()}}
}
\concept{adding badges}
