% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_codemeta.R
\name{add_github_actions_codemeta}
\alias{add_github_actions_codemeta}
\title{Setup GitHub Actions to update codemeta.json}
\usage{
add_github_actions_codemeta(open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a configuration file (\code{.yaml}) to setup GitHub Actions
to update the \code{codemeta.json}.
This workflow is derived
from \url{https://github.com/r-lib/actions/tree/v2-branch/examples}.
This file will be written as \code{.github/workflows/update-citation-cff.yaml}.

This function also create the \code{codemeta.json} using the package \link{codemetar}.
}
\examples{
\dontrun{
add_github_actions_codemeta()
}
}
\seealso{
Other development functions: 
\code{\link{add_dependencies}()},
\code{\link{add_github_actions_check}()},
\code{\link{add_github_actions_citation}()},
\code{\link{add_github_actions_codecov}()},
\code{\link{add_github_actions_document}()},
\code{\link{add_github_actions_pkgdown}()},
\code{\link{add_github_actions_render}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
