% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_pkgdown.R
\name{add_github_actions_pkgdown}
\alias{add_github_actions_pkgdown}
\title{Setup GitHub Actions to build and deploy package website}
\usage{
add_github_actions_pkgdown(open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{open}{a logical value. If \code{TRUE} the file is opened in the editor.
Default is \code{FALSE}.}

\item{overwrite}{a logical value. If this file already exists and
\code{overwrite = TRUE}, its content will be erased.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function creates a configuration file (\code{.yaml}) to setup GitHub Actions
to automatically build and deploy the website using \code{\link[pkgdown:pkgdown-package]{pkgdown::pkgdown()}}.
This file will be added (from a template) in the folder
\code{.github/workflows/pkgdown.yaml}. An additional (empty) file \verb{_pkgdown.yaml}
will also be written: it can be use to customize the package website.
}
\examples{
\dontrun{
add_github_actions_pkgdown()
}
}
\seealso{
Other development functions: 
\code{\link{add_dependencies}()},
\code{\link{add_github_actions_check}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
