% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_render.R
\name{add_github_actions_render}
\alias{add_github_actions_render}
\title{Setup GitHub Actions to render README}
\usage{
add_github_actions_render(open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a configuration file (\code{.yaml}) to setup GitHub Actions
to automatically knit the \code{README.Rmd} after a push. This workflow will be
triggered only if the \code{README.Rmd} has been modified since the last commit.
This workflow is derived
from \url{https://github.com/r-lib/actions/tree/v2-branch/examples}.
This file will be written as \code{.github/workflows/render-README.yaml}.
}
\examples{
\dontrun{
add_github_actions_render()
}
}
\seealso{
Other development functions: 
\code{\link{add_dependencies}()},
\code{\link{add_github_actions_check}()},
\code{\link{add_github_actions_codecov}()},
\code{\link{add_github_actions_pkgdown}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
