% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_testthat.R
\name{add_testthat}
\alias{add_testthat}
\title{Initialize units tests}
\usage{
add_testthat()
}
\value{
No return value.
}
\description{
This function initializes units tests settings by running
\code{\link[usethis:use_testthat]{usethis::use_testthat()}} and by adding an example units tests file
\code{tests/testthat/test-demo.R}. The sample file will test a demo function
created in \code{R/fun-demo.R}.
}
\examples{
\dontrun{
add_testthat()
}
}
\seealso{
Other create files:
\code{\link[=add_citation]{add_citation()}},
\code{\link[=add_code_of_conduct]{add_code_of_conduct()}},
\code{\link[=add_codeowners]{add_codeowners()}},
\code{\link[=add_compendium]{add_compendium()}},
\code{\link[=add_contributing]{add_contributing()}},
\code{\link[=add_description]{add_description()}},
\code{\link[=add_dockerfile]{add_dockerfile()}},
\code{\link[=add_license]{add_license()}},
\code{\link[=add_makefile]{add_makefile()}},
\code{\link[=add_package_doc]{add_package_doc()}},
\code{\link[=add_readme_rmd]{add_readme_rmd()}},
\code{\link[=add_renv]{add_renv()}},
\code{\link[=add_vignette]{add_vignette()}}
}
\concept{create files}
