% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vignette.R
\name{add_vignette}
\alias{add_vignette}
\title{Create a vignette document}
\usage{
add_vignette(
  filename = NULL,
  title = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{filename}{A character of length 1. The name of the \code{.Rmd} file to be
created. If \code{NULL} (default ) the \code{.Rmd} file will be named \code{pkg.Rmd}
where \code{pkg} is your package name.}

\item{title}{A character of length 1. The title of the vignette. If \code{NULL}
(default) the title will be \verb{Get started}.}

\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function adds a vignette in the folder \verb{vignettes/}. It also adds
dependencies \href{https://yihui.org/knitr/}{\code{knitr}} and
\href{https://rmarkdown.rstudio.com/}{\code{rmarkdown}} in the
field \code{Suggests} of the \code{DESCRIPTION} file (if not already present in
fields \code{Imports}).
}
\examples{
\dontrun{
## Default vignette ----
add_vignette()

## Default vignette ----
add_vignette(filename = "pkg", title = "Get started")
}
}
\seealso{
Other create files:
\code{\link[=add_citation]{add_citation()}},
\code{\link[=add_code_of_conduct]{add_code_of_conduct()}},
\code{\link[=add_codeowners]{add_codeowners()}},
\code{\link[=add_compendium]{add_compendium()}},
\code{\link[=add_contributing]{add_contributing()}},
\code{\link[=add_description]{add_description()}},
\code{\link[=add_dockerfile]{add_dockerfile()}},
\code{\link[=add_license]{add_license()}},
\code{\link[=add_makefile]{add_makefile()}},
\code{\link[=add_package_doc]{add_package_doc()}},
\code{\link[=add_readme_rmd]{add_readme_rmd()}},
\code{\link[=add_renv]{add_renv()}},
\code{\link[=add_testthat]{add_testthat()}}
}
\concept{create files}
