% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_functions.R
\name{get_all_functions}
\alias{get_all_functions}
\title{List all functions in the package}
\usage{
get_all_functions()
}
\value{
A list of two vectors:
\itemize{
\item \code{external}, a vector of exported functions name;
\item \code{internal}, a vector of internal functions name.
}
}
\description{
This function returns a list of all the functions (exported and internal)
available with the project. As this function scans the \code{NAMESPACE} and the
\verb{R/} folder, it is recommended to:
\itemize{
\item store all the functions in the \verb{R/} folder
\item run \code{\link[devtools:document]{devtools::document()}} before to update the \code{NAMESPACE}
}
}
\examples{
\dontrun{
## Update NAMESPACE ----
devtools::document()

## List all implemented functions ----
get_all_functions()
}
}
\seealso{
Other utilities functions:
\code{\link[=get_all_dependencies]{get_all_dependencies()}},
\code{\link[=get_available_gh_actions]{get_available_gh_actions()}},
\code{\link[=get_available_issue_templates]{get_available_issue_templates()}},
\code{\link[=get_available_licenses]{get_available_licenses()}},
\code{\link[=get_minimal_r_version]{get_minimal_r_version()}}
}
\concept{utilities functions}
