% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_licenses.R
\name{get_available_licenses}
\alias{get_available_licenses}
\title{List all available licenses}
\usage{
get_available_licenses()
}
\value{
A \code{data.frame} with the following two variables:
\itemize{
\item \code{tag}, the license name to be used with \code{\link[=add_license]{add_license()}};
\item \code{url}, the URL of the license description.
}
}
\description{
This function returns a list of all available licenses. This is particularly
useful to get the right spelling of the license to be passed to
\code{\link[=create_new_package]{create_new_package()}}, \code{\link[=create_new_compendium]{create_new_compendium()}}, or \code{\link[=add_license]{add_license()}}.
}
\examples{
get_available_licenses()
}
\seealso{
Other utilities functions:
\code{\link[=get_all_dependencies]{get_all_dependencies()}},
\code{\link[=get_all_functions]{get_all_functions()}},
\code{\link[=get_available_gh_actions]{get_available_gh_actions()}},
\code{\link[=get_available_issue_templates]{get_available_issue_templates()}},
\code{\link[=get_minimal_r_version]{get_minimal_r_version()}}
}
\concept{utilities functions}
