% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get Required Minimal R Version}
\usage{
get_dependencies(pkg)
}
\arguments{
\item{pkg}{...}
}
\value{
...
}
\description{
This function...
}
\examples{
\dontrun{
##
}
}
