% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_functions.R
\name{get_pkg_functions}
\alias{get_pkg_functions}
\title{List all functions in the package}
\usage{
get_pkg_functions()
}
\value{
A vector of internal and external (exported) functions of the packages.
Exported functions are noted \code{pkg::fun()} while internal functions are noted
\code{pkg:::fun()}.
}
\description{
This function returns a list of all the functions (exported and internal)
available with the package. As this function scans the \strong{NAMESPACE} it is
recommended to update this file with \code{\link[devtools:document]{devtools::document()}}.
}
\examples{
\dontrun{
devtools::document()
get_pkg_functions()
}
}
\concept{utilitaries functions}
