% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rd_families.R
\name{get_rd_families}
\alias{get_rd_families}
\title{List all Rd functions families}
\usage{
get_rd_families()
}
\value{
A vector of Rd function families.
}
\description{
This function parses all R functions to detect \verb{@family} tag. For internal
purpose.
}
\examples{
\dontrun{
get_rd_families()
}
}
