% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refresh a package/research compendium}
\usage{
refresh(compendium = NULL, make = FALSE, check = TRUE, quiet = FALSE)
}
\arguments{
\item{compendium}{(optional) the name of the folder to recursively detect
dependencies to be added in the \code{Imports} field of \strong{DESCRIPTION} file.
Default is \code{compendium = NULL} (but \strong{R/}, \strong{NAMESPACE}, and \verb{@examples}
are still inspected).}

\item{make}{a logical value. If \code{TRUE} the Make-like R file \code{make.R} is
sourced. Only for research compendium. Default is \code{FALSE}.}

\item{check}{a logical value. If \code{TRUE} (default) package integrity is
checked using \code{\link[devtools:check]{devtools::check()}}.}

\item{quiet}{a logical value. If \code{TRUE} (default) message are deleted.}
}
\description{
This function refreshes a package/research compendium. It will:
\itemize{
\item Update \code{.Rd} files and \code{NAMESPACE}, using \code{\link[devtools:document]{devtools::document()}};
\item Update external packages (in \code{DESCRIPTION} file);
\item Update badges in \code{README.Rmd} (if already present);
\item Re-knitr the \code{README.Rmd}, using \code{\link[rmarkdown:render]{rmarkdown::render()}};
\item Check package integrity with \code{\link[devtools:check]{devtools::check()}};
\item Run analysis by sourcing \code{make.R} (only for compendium).
}
}
\examples{
\dontrun{
library(rcompendium)

## Create an R package ----
new_package()

## Start developing functions ----
## ...

## Update package (documentation, dependencies, README, check) ----
refresh()
}
}
\seealso{
Other setup functions: 
\code{\link{new_compendium}()},
\code{\link{new_package}()},
\code{\link{set_credentials}()}
}
\concept{setup functions}
