% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_citation.R
\name{add_citation}
\alias{add_citation}
\title{Create a CITATION file}
\usage{
add_citation(
  given = NULL,
  family = NULL,
  github = NULL,
  organisation = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{given}{(optional) user given name (i.e. "Nicolas")}

\item{family}{(optional) user family name (i.e. "Casajus")}

\item{github}{(optional) user GitHub pseudo/organization (i.e. "ahasverus")}

\item{organisation}{a character. The name of the GITHUB organisation to
host the package. If \code{NULL} it uses the GITHUB account.}

\item{open}{a logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{a logical value. If a \emph{CITATION} is already present and
\code{overwrite = TRUE}, this file will be erased and replaced.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function creates a \emph{CITATION} file in the folder \emph{inst/}. This file
contains a BiBTeX entry to cite the package as a manual.
}
\examples{
\dontrun{
add_citation()
readCitationFile("inst/CITATION")
citation("pkg")    # If you have installed your package <pkg>
}
}
\seealso{
Other create files: 
\code{\link{add_description}()},
\code{\link{add_license}()},
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()}
}
\concept{create files}
