% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_code_of_conduct.R
\name{add_code_of_conduct}
\alias{add_code_of_conduct}
\title{Add code of conduct}
\usage{
add_code_of_conduct(
  email = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{email}{A character of length 1. The email address of the project
maintainer.}

\item{open}{A logical value. If \code{TRUE} (default) the \code{CONTRIBUTING.md} file
is opened in the editor.}

\item{overwrite}{A logical value. If files are already present and
\code{overwrite = TRUE}, they will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a \code{CODE_OF_CONDUCT.md} file adapted from the
Contributor Covenant, version 2.1 available at
\url{https://www.contributor-covenant.org/version/2/1/code_of_conduct.html}.
}
\examples{
\dontrun{
add_code_of_conduct()
}
}
\seealso{
Other create files: 
\code{\link{add_citation}()},
\code{\link{add_compendium}()},
\code{\link{add_contributing}()},
\code{\link{add_description}()},
\code{\link{add_dockerfile}()},
\code{\link{add_license}()},
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()},
\code{\link{add_renv}()},
\code{\link{add_testthat}()},
\code{\link{add_vignette}()}
}
\concept{create files}
