% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_codecov_badge.R
\name{add_codecov_badge}
\alias{add_codecov_badge}
\title{Add a Codecov badge}
\usage{
add_codecov_badge(organisation = NULL, quiet = FALSE)
}
\arguments{
\item{organisation}{a character of length 1

The name of the GitHub organisation to host the package. If \code{NULL}
(default) the GitHub account will be used.}

\item{quiet}{a logical value

If \code{TRUE} messages are deleted. Default is \code{FALSE}.}
}
\value{
A Markdown badge expression
}
\description{
This function adds a Codecov badge to the \code{README.Rmd}, i.e. the percentage
of code cover by units tests. This percentage is computed by codecov.io
service.
}
\examples{
\dontrun{
add_codecov_badge()
}
}
\seealso{
Other adding badges: 
\code{\link{add_cran_badge}()},
\code{\link{add_dependencies_badge}()},
\code{\link{add_github_actions_check_badge}()},
\code{\link{add_github_actions_codecov_badge}()},
\code{\link{add_github_actions_pkgdown_badge}()},
\code{\link{add_license_badge}()},
\code{\link{add_lifecycle_badge}()},
\code{\link{add_repostatus_badge}()}
}
\concept{adding badges}
