% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_codeowners.R
\name{add_codeowners}
\alias{add_codeowners}
\title{Create a CODEOWNERS file}
\usage{
add_codeowners(open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a \code{CODEOWNERS} file in the folder \verb{.github/}. This
file is used to define individual that is responsible for code in the
repository.
}
\examples{
\dontrun{
add_codeowners()
}
}
\seealso{
Other create files: 
\code{\link{add_citation}()},
\code{\link{add_code_of_conduct}()},
\code{\link{add_compendium}()},
\code{\link{add_contributing}()},
\code{\link{add_description}()},
\code{\link{add_dockerfile}()},
\code{\link{add_license}()},
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()},
\code{\link{add_renv}()},
\code{\link{add_testthat}()},
\code{\link{add_vignette}()}
}
\concept{create files}
