% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_compendium.R
\name{add_compendium}
\alias{add_compendium}
\title{Create compendium folders}
\usage{
add_compendium(compendium = ".")
}
\arguments{
\item{compendium}{a character of length 1

By default, compendium folders are created at the root of the project.
User can change their location with this argument. For instance, if
\code{compendium = 'analysis'}, compendium folders will be created inside the
directory \verb{analysis/}.}
}
\value{
None
}
\description{
This function creates the following additional folders \verb{data/}, \verb{rscripts/},
\verb{outputs/}, \verb{figures/}, and \verb{paper/}. Each folder has a README to provide
instructions. The argument compendium allows user to choose if these folders
are created at the root of the project (default) or nested inside a parent
directory. All theses folders area added to \code{.Rbuildignore}.
}
\examples{
\dontrun{
add_compendium()
}
}
\seealso{
Other create files: 
\code{\link{add_citation}()},
\code{\link{add_description}()},
\code{\link{add_license}()},
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()},
\code{\link{add_testthat}()},
\code{\link{add_vignette}()}
}
\concept{create files}
