% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cran_badge.R
\name{add_cran_badge}
\alias{add_cran_badge}
\title{Add a CRAN status badge}
\usage{
add_cran_badge(quiet = FALSE)
}
\arguments{
\item{quiet}{a logical value

If \code{TRUE} messages are deleted. Default is \code{FALSE}.}
}
\value{
A Markdown badge expression
}
\description{
This function adds a CRAN status badge to the \code{README.Rmd}. If the package
is not hosted on the CRAN the badge will indicate
\emph{not published on the CRAN}.
}
\examples{
\dontrun{
add_cran_badge()
}
}
\seealso{
Other adding badges: 
\code{\link{add_codecov_badge}()},
\code{\link{add_dependencies_badge}()},
\code{\link{add_github_actions_check_badge}()},
\code{\link{add_github_actions_codecov_badge}()},
\code{\link{add_github_actions_pkgdown_badge}()},
\code{\link{add_license_badge}()},
\code{\link{add_lifecycle_badge}()},
\code{\link{add_repostatus_badge}()}
}
\concept{adding badges}
