% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependencies.R
\name{add_dependencies}
\alias{add_dependencies}
\title{Add Dependencies in DESCRIPTION File}
\usage{
add_dependencies(path = ".", import = NULL, suggest = "vignettes")
}
\arguments{
\item{path}{the path to the package/research compendium (must have a
\code{DESCRIPTION} file) in which dependencies are recursively detected.}

\item{import}{(optional) the name of the folder to recursively detect
dependencies to be added in the \strong{Imports} field of \code{DESCRIPTION} file.
Default is \code{import = NULL} (but \verb{R/}, \code{NAMESPACE}, and \verb{@examples} are still
inspected).}

\item{suggest}{the name of the folder to recursively detect dependencies to
be added in the \strong{Suggests} field of \code{DESCRIPTION} file.
Default is \code{suggest = "vignettes"}.}
}
\description{
This function detects dependencies used in \verb{R/}, \code{NAMESPACE}, and \verb{@examples}
and automatically adds these dependencies in \strong{Imports} section of the
\code{DESCRIPTION} file.
}
\details{
In the \code{NAMESPACE} file it detects dependencies mentionned as \code{import(pkg)}
and \code{importFrom(pkg,fun)}.

In the \verb{R/} folder it detects functions called as \code{pkg::fun()} in the code
of each R files. In \verb{@examples} section it also detects packages attached
using \code{library(pkg)} or \code{require(pkg)}.

The folder \verb{vignettes/} is also inspected and detected dependencies
(\code{pkg::fun()}, \code{library(pkg)} or \code{require(pkg)}) are added in the section
\strong{Suggests} of the \code{DESCRIPTION} file.

If the project is a research compendium user can also inspect one
additional folder (i.e. \verb{analysis/}) with the argument \code{import} to detected
dependencies to be added in the section \strong{Imports} of the \code{DESCRIPTION}
file. The detection process is the same as the one used for \verb{vignettes/}.
}
\examples{
\dontrun{
add_dependencies()
}
}
