% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_description}
\alias{add_description}
\title{Create a DESCRIPTION File}
\usage{
add_description(open = TRUE, overwrite = FALSE)
}
\arguments{
\item{open}{a logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{a logical value. If a \code{DESCRIPTION} is already present and
\code{overwrite = TRUE}, this file will be erased and replaced.}
}
\description{
This function creates a DESCRIPTION file at the root of the project based on
a template.
}
\examples{
\dontrun{
add_description()
}
}
\seealso{
Other create files: 
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()}
}
\concept{create files}
