% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_action.R
\name{add_github_action}
\alias{add_github_action}
\title{Set up a GitHub Action workflow}
\usage{
add_github_action(name, open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{name}{A character of length 1. The name of the GitHub Action to set up.
Run \code{\link[=get_available_gh_actions]{get_available_gh_actions()}} to list available GitHub Actions.}

\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function sets up a continuous integration (CI) for an R project hosted
on GitHub. It creates a configuration file (\code{.yaml}) to setup a specific
GitHub Action (see below).

Available workflows are derived from
\url{https://github.com/r-lib/actions/tree/v2-branch/examples} and are
hosted on a different repository
\url{https://github.com/FRBCesab/r-templates/tree/main/actions}.

Configuration files will be written in \verb{.github/workflows/}.
}
\details{
Here is a list of commonly used GitHub Actions that will be triggered after
each push or pull request:
\itemize{
\item \code{R-CMD-check}: check the package integrity on three major operating systems
(Ubuntu, macOS, and Windows) using the latest release of R. The package is
also checked on Ubuntu (latest version) using the development and previous
versions of R.
\item \code{test-coverage}: compute and report the code coverage (i.e. the proportion
of R code covered by unit tests) to \url{https://about.codecov.io/}.
\item \code{pkgdown}: build and publish the \code{pkgdown} website of the package on a
dedicated \strong{gh-pages} branch. The repository may need further configuration
(see \code{\link[usethis:use_pkgdown]{usethis::use_pkgdown_github_pages()}}).
\item \code{check-format}: set up the Air formatter for R and check the code format.
This action will fail if any files need to be reformatted with Air.
\item \code{render-README}: render the \code{README.Rmd} only if it has been modified.
This action will commit and push the updated \code{README.md} to the main branch.
User need to fetch the new version.
\item \code{update-Rd-files}: update the package documentation files (\code{Rd} files and
\code{NAMESPACE}). This action will commit and push the updated files to the main
branch. User need to fetch the new version.
\item \code{update-citation-cff}: update the \code{CITATION.cff}, a human- and
machine-readable citation information file. This action is triggered if
the \code{DESCRIPTION} file and/or the \code{inst/CITATION} file is modified. This
action will commit and push the updated \code{CITATION.cff} to the main branch.
User need to fetch the new version.
\item \code{update-codemeta}: update the \code{codemeta.json}, a standardized metadata
file for software (not only R packages). This action is triggered if
the \code{DESCRIPTION} file and/or the \code{inst/CITATION} file and/or the \code{README.md}
is modified. This action will commit and push the updated \code{codemeta.json} to
the main branch. User need to fetch the new version.
}

If one of these GitHub Actions is added to the project, this function will
also create a \code{dependabot.yaml} file in \verb{.github/}. This action will be
triggered once a week to check for dependency updates (used by any GitHub
Action of the repository). If an update is available, this bot will open a
Pull Request and user will be invited to review changes.
}
\examples{
\dontrun{
add_github_action(name = "R-CMD-check")
}
}
\seealso{
Other development functions: 
\code{\link{add_dependabot}()},
\code{\link{add_dependencies}()},
\code{\link{add_issue_template}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
