% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_check.R
\name{add_github_actions_check}
\alias{add_github_actions_check}
\title{Setup GitHub Actions to check package}
\usage{
add_github_actions_check(open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a configuration file (\code{.yaml}) to setup GitHub Actions
to check the package.
This workflow is derived
from \url{https://github.com/r-lib/actions/tree/v2-branch/examples}.
This file will be written as \code{.github/workflows/R-CMD-check.yaml}.
}
\details{
This workflow runs \verb{R CMD check} on the three major operating systems
(Ubuntu, macOS, and Windows) using the latest release of R. The package is
also checked on Ubuntu (latest version) using the development and previous
versions of R.
}
\examples{
\dontrun{
add_github_actions_check()
}
}
\seealso{
Other development functions: 
\code{\link{add_dependencies}()},
\code{\link{add_github_actions_codecov}()},
\code{\link{add_github_actions_pkgdown}()},
\code{\link{add_github_actions_render}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
