% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_github_actions_pkgdown_badge.R
\name{add_github_actions_pkgdown_badge}
\alias{add_github_actions_pkgdown_badge}
\title{Add a Website Deployment badge}
\usage{
add_github_actions_pkgdown_badge(
  github = NULL,
  organisation = NULL,
  quiet = FALSE
)
}
\arguments{
\item{github}{user GitHub account where the project is hosted. If \code{NULL}
(default) the function will try to get this value by reading the
\code{.Rprofile} file (unless \code{!is.null(organisation)}).}

\item{organisation}{(alternative) name of the GitHub organisation where the
project is hosted.  If \code{NULL} (default) the GitHub account \code{github} will
be used.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function adds a Website Deployment badge to the \code{README.Rmd}. This
function must be run after \code{\link[=add_github_actions_pkgdown]{add_github_actions_pkgdown()}} which will setup
GitHub Actions (website deployment).
}
\examples{
\dontrun{
add_github_actions_pkgdown_badge()
}
}
\seealso{
Other adding badges: 
\code{\link{add_cran_badge}()},
\code{\link{add_dependencies_badge}()},
\code{\link{add_github_actions_check_badge}()},
\code{\link{add_license_badge}()},
\code{\link{add_lifecycle_badge}()},
\code{\link{add_repostatus_badge}()}
}
\concept{adding badges}
