% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_issue_template.R
\name{add_issue_template}
\alias{add_issue_template}
\title{Set up an Issue Template file}
\usage{
add_issue_template(name, open = FALSE, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{name}{A character of length 1. The name of the Issue Template to add.
Run \code{\link[=get_available_issue_templates]{get_available_issue_templates()}} to list available Issue Templates.}

\item{open}{A logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{A logical value. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced. Default is \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates an Issue template file (\code{md}) in the directory
\code{.github/ISSUE_TEMPLATE}. These files preformat a GitHub Issue. Contributors
can use these templates when they open new issues. For instance, you can
format issue related to bug report, feature request, etc.
}
\examples{
\dontrun{
add_issue_template(name = "feature_request")
}
}
\seealso{
Other development functions: 
\code{\link{add_dependabot}()},
\code{\link{add_dependencies}()},
\code{\link{add_github_action}()},
\code{\link{add_r_depend}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
