% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_license.R
\name{add_license}
\alias{add_license}
\title{Add a LICENSE file}
\usage{
add_license(license = NULL, given = NULL, family = NULL, quiet = FALSE)
}
\arguments{
\item{license}{A character of length 1. The license name.
Run \code{\link[=get_licenses]{get_licenses()}}) to select an appropriate one.}

\item{given}{a \code{character} of length 1. The given name of the user
(considered as the maintainer and code owner of the project).}

\item{family}{a \code{character} of length 1. The family name of the user
(considered as the maintainer and code owner of the project).}

\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function adds a license to the project. It will add the license name
in the \code{License} field of the \code{DESCRIPTION} file and write the content of
the license in a \code{LICENSE.md} file.
}
\examples{
\dontrun{
add_license(license = "MIT")
add_license(license = "GPL (>= 2)")
}
}
\seealso{
Other create files:
\code{\link[=add_citation]{add_citation()}},
\code{\link[=add_code_of_conduct]{add_code_of_conduct()}},
\code{\link[=add_codeowners]{add_codeowners()}},
\code{\link[=add_compendium]{add_compendium()}},
\code{\link[=add_contributing]{add_contributing()}},
\code{\link[=add_description]{add_description()}},
\code{\link[=add_dockerfile]{add_dockerfile()}},
\code{\link[=add_makefile]{add_makefile()}},
\code{\link[=add_package_doc]{add_package_doc()}},
\code{\link[=add_readme_rmd]{add_readme_rmd()}},
\code{\link[=add_renv]{add_renv()}},
\code{\link[=add_testthat]{add_testthat()}},
\code{\link[=add_vignette]{add_vignette()}}
}
\concept{create files}
