% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_license_badge.R
\name{add_license_badge}
\alias{add_license_badge}
\title{Add a license badge}
\usage{
add_license_badge(quiet = FALSE)
}
\arguments{
\item{quiet}{a logical value

If \code{TRUE} messages are deleted. Default is \code{FALSE}.}
}
\value{
A Markdown badge expression
}
\description{
This function adds or updates the license badge to the \code{README.Rmd}. This
function reads the \code{License} field of the \code{DESCRIPTION} file. Be sure to
previously run \code{\link[=add_license]{add_license()}} function.
}
\examples{
\dontrun{
add_license_badge()
}
}
\seealso{
Other adding badges: 
\code{\link{add_codecov_badge}()},
\code{\link{add_cran_badge}()},
\code{\link{add_dependencies_badge}()},
\code{\link{add_github_actions_check_badge}()},
\code{\link{add_github_actions_codecov_badge}()},
\code{\link{add_github_actions_pkgdown_badge}()},
\code{\link{add_lifecycle_badge}()},
\code{\link{add_repostatus_badge}()}
}
\concept{adding badges}
