% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_makefile.R
\name{add_makefile}
\alias{add_makefile}
\title{Create a Make-like R File}
\usage{
add_makefile(open = TRUE, overwrite = FALSE)
}
\arguments{
\item{open}{a logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{a logical value. If a \code{make.R} is already present and
\code{overwrite = TRUE}, this file will be erased and replaced.}
}
\description{
This function creates a Make-like R file (\code{make.R}) at the root of the
project based on a template.
}
\examples{
\dontrun{
add_makefile()
}
}
\seealso{
Other create files: 
\code{\link{add_description}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()}
}
\concept{create files}
