% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_r_depend.R
\name{add_r_depend}
\alias{add_r_depend}
\title{Add minimal R version to DESCRIPTION}
\usage{
add_r_depend()
}
\value{
No return value.
}
\description{
This function adds the minimal R version to the \code{Depends} field of the
\code{DESCRIPTION} file. This version corresponds to the higher version of R
among all dependencies. If no dependencies mentions minimal R version, the
\code{DESCRIPTION} is not modified.
}
\examples{
\dontrun{
add_r_depend()
}
}
\seealso{
Other development functions: 
\code{\link{add_dependencies}()},
\code{\link{add_github_actions_check}()},
\code{\link{add_github_actions_codecov}()},
\code{\link{add_github_actions_pkgdown}()},
\code{\link{add_github_actions_render}()},
\code{\link{add_to_buildignore}()},
\code{\link{add_to_gitignore}()}
}
\concept{development functions}
