% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_readme_rmd.R
\name{add_readme_rmd}
\alias{add_readme_rmd}
\title{Create a README file}
\usage{
add_readme_rmd(
  type = "package",
  given = NULL,
  family = NULL,
  organisation = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{type}{a character. If \code{package} (default) a GitHub \code{README.Rmd}
specific to an R package will be created. If \code{compendium} a GitHub
\code{README.Rmd} specific to a research compendium will be created.}

\item{given}{(optional) given name of the package maintainer}

\item{family}{(optional) family name of the package maintainer}

\item{organisation}{a character. The name of the GITHUB organisation to
host the package. If \code{NULL} the GITHUB account will be used.}

\item{open}{a logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{a logical value. If a \code{README.Rmd} is already present and
\code{overwrite = TRUE}, this file will be erased and replaced.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function creates a \code{README.Rmd} file at the root of the project based
on a template. Once edited user needs to knit it into a \code{README.Rmd} (or use
the function \code{\link[=refresh]{refresh()}}.
}
\examples{
\dontrun{
add_readme_rmd(type = "package")
}
}
\seealso{
Other create files: 
\code{\link{add_citation}()},
\code{\link{add_description}()},
\code{\link{add_license}()},
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_testthat}()},
\code{\link{add_vignette}()}
}
\concept{create files}
