% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_renv.R
\name{add_renv}
\alias{add_renv}
\title{Initialize renv}
\usage{
add_renv(quiet = FALSE)
}
\arguments{
\item{quiet}{A logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\value{
No return value.
}
\description{
This function initializes an \code{renv} environment for the project by running
\code{\link[renv:init]{renv::init()}}. See \url{https://rstudio.github.io/renv/} for further detail.
}
\examples{
\dontrun{
add_renv()
}
}
\seealso{
Other create files: 
\code{\link{add_citation}()},
\code{\link{add_compendium}()},
\code{\link{add_description}()},
\code{\link{add_dockerfile}()},
\code{\link{add_license}()},
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()},
\code{\link{add_testthat}()},
\code{\link{add_vignette}()}
}
\concept{create files}
