% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_buildignore.R
\name{add_to_buildignore}
\alias{add_to_buildignore}
\title{Add to .Rbuildignore File}
\usage{
add_to_buildignore(x, open = FALSE)
}
\arguments{
\item{x}{a character of one or several files/directories names to add to the
\code{.Rbuildignore}. This argument is mandatory.}

\item{open}{a logical value. If \code{TRUE} the file is opened in the editor.
Default is \code{FALSE}.}
}
\description{
This function adds files/folders to the \code{.Rbuildignore} file. If a
\code{.Rbuildignore} is already present, files to be ignored are just added to
this file. Otherwise a new file is created.
}
