% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_gitignore.R
\name{add_to_gitignore}
\alias{add_to_gitignore}
\title{Create a .gitignore File}
\usage{
add_to_gitignore(x, open = FALSE)
}
\arguments{
\item{x}{a character of one or several files/directories names to add to the
\code{.gitignore}.}

\item{open}{a logical value. If \code{TRUE} the file is opened in the editor.
Default is \code{FALSE}.}
}
\description{
This function creates a \code{.gitignore} file at the root of the project based on
a template (specific to R). If a \code{.gitignore} is already present, files to
be untracked by \strong{git} are just added to this file.
}
