% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vignette.R
\name{add_vignette}
\alias{add_vignette}
\title{Create a vignette document}
\usage{
add_vignette(
  filename = NULL,
  title = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{filename}{the name of the \code{.Rmd} file to be created. If \code{NULL} the
\code{.Rmd} file will be named \code{pkg.Rmd} where \code{pkg} is your package name (if
does not exist).}

\item{title}{the title of the vignette. If \code{NULL} the title will be
\verb{Introduction to pkg} (where \code{pkg} is your package name).}

\item{open}{a logical value. If \code{TRUE} (default) the file is opened in the
editor.}

\item{overwrite}{a logical value. If a file is already present and
\code{overwrite = TRUE}, it will be erased and replaced.}

\item{quiet}{a logical value. If \code{TRUE} messages are deleted. Default is
\code{FALSE}.}
}
\description{
This function adds a vignette in the folder \verb{vignettes/}. It also adds
dependencies \code{\link[knitr:knitr-package]{knitr::knitr()}} and \code{\link[rmarkdown:rmarkdown-package]{rmarkdown::rmarkdown()}} in the
field \code{Suggests} of the \code{DESCRIPTION} file (if not already present in
fields \code{Imports} and \code{Suggests}).
}
\examples{
\dontrun{
## Default vignette ----
add_vignette()

## Custom vignette ----
add_vignette(filename = "get_started", title = "Get started")
}
}
\seealso{
Other create files: 
\code{\link{add_citation}()},
\code{\link{add_description}()},
\code{\link{add_license}()},
\code{\link{add_makefile}()},
\code{\link{add_package_doc}()},
\code{\link{add_readme_rmd}()},
\code{\link{add_testthat}()}
}
\concept{create files}
