% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_functions.R
\name{get_all_functions}
\alias{get_all_functions}
\title{List all functions in the package}
\usage{
get_all_functions()
}
\value{
A list of two vectors:
\itemize{
\item \code{external}, a vector of exported functions name (\code{pkg::fun()});
\item \code{internal}, a vector of internal functions name (\code{pkg:::fun()}).
}
}
\description{
This function returns a list of all the functions (exported and internal)
available with the package. As this function scans the \strong{NAMESPACE} and the
\strong{R/} folder, it is recommended to run \code{\link[devtools:document]{devtools::document()}} before.
}
\examples{
\dontrun{
devtools::document()
get_all_functions()
}
}
\seealso{
Other utilities functions: 
\code{\link{get_all_dependencies}()},
\code{\link{get_licenses}()},
\code{\link{get_minimal_r_version}()},
\code{\link{get_rd_families}()}
}
\concept{utilities functions}
