% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_gh_actions.R
\name{get_available_gh_actions}
\alias{get_available_gh_actions}
\title{List available GitHub Actions}
\usage{
get_available_gh_actions()
}
\value{
A \code{character} vector with the name of the available GitHub Actions in
\code{rcompendium}.
}
\description{
This function returns a list of the name of all available GitHub Actions.
This is particularly useful to get the right spelling of the GitHub Action
to be passed to \code{\link[=add_github_action]{add_github_action()}}.
}
\examples{
get_available_gh_actions()
}
\seealso{
Other utilities functions: 
\code{\link{get_all_dependencies}()},
\code{\link{get_all_functions}()},
\code{\link{get_licenses}()},
\code{\link{get_minimal_r_version}()}
}
\concept{utilities functions}
