% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_issue_template.R
\name{get_available_issue_templates}
\alias{get_available_issue_templates}
\title{List available ISSUE TEMPLATE}
\usage{
get_available_issue_templates()
}
\value{
A \code{character} vector with the name of the available Issue Templates
in \code{rcompendium}.
}
\description{
This function returns a list of the name of all available Issue Templates.
This is particularly useful to get the right spelling of the Issue Templates
to be passed to \code{\link[=add_issue_template]{add_issue_template()}}.
}
\examples{
\dontrun{
get_available_issue_templates()
}
}
\seealso{
Other utilities functions:
\code{\link[=get_all_dependencies]{get_all_dependencies()}},
\code{\link[=get_all_functions]{get_all_functions()}},
\code{\link[=get_available_gh_actions]{get_available_gh_actions()}},
\code{\link[=get_licenses]{get_licenses()}},
\code{\link[=get_minimal_r_version]{get_minimal_r_version()}}
}
\concept{utilities functions}
