% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_licenses.R
\name{get_licenses}
\alias{get_licenses}
\title{List all available licenses}
\usage{
get_licenses()
}
\value{
A data frame with the following two variables:
\itemize{
\item \code{tag}, the license name to be used with \code{\link[=add_license]{add_license()}};
\item \code{url}, the URL of the license description.
}
}
\description{
This function returns a list of all available licenses. This is particularly
useful to get the right spelling of the license to be passed to
\code{\link[=new_package]{new_package()}}, \code{\link[=new_compendium]{new_compendium()}}, or \code{\link[=add_license]{add_license()}}.
}
\examples{
\dontrun{
get_licenses()

##                      tag                                            url
## 1                    MIT        https://choosealicense.com/licenses/mit
## 2                    CC0    https://choosealicense.com/licenses/cc0-1.0
## 3              CC BY 4.0  https://choosealicense.com/licenses/cc-by-4.0
## 4                  GPL-2    https://choosealicense.com/licenses/gpl-2.0
## 5                  GPL-3    https://choosealicense.com/licenses/gpl-3.0
## 6             GPL (>= 2)    https://choosealicense.com/licenses/gpl-2.0
## 7             GPL (>= 3)    https://choosealicense.com/licenses/gpl-3.0
## 8               LGPL-2.1   https://choosealicense.com/licenses/lgpl-2.1
## 9                 LGPL-3   https://choosealicense.com/licenses/lgpl-3.0
## 10         LGPL (>= 2.1)   https://choosealicense.com/licenses/lgpl-2.1
## 11           LGPL (>= 3)   https://choosealicense.com/licenses/lgpl-3.0
## 12                AGPL-3   https://choosealicense.com/licenses/agpl-3.0
## 13           AGPL (>= 3)   https://choosealicense.com/licenses/agpl-3.0
## 14 Apache License (== 2) https://choosealicense.com/licenses/apache-2.0
## 15 Apache License (>= 2) https://choosealicense.com/licenses/apache-2.0
}
}
\seealso{
Other utilities functions: 
\code{\link{get_all_dependencies}()},
\code{\link{get_all_functions}()},
\code{\link{get_minimal_r_version}()}
}
\concept{utilities functions}
