% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_licenses.R
\name{get_licenses}
\alias{get_licenses}
\title{List all available licenses}
\usage{
get_licenses()
}
\value{
A \code{data.frame} with the following two variables:
\itemize{
\item \code{tag}, the license name to be used with \code{\link[=add_license]{add_license()}};
\item \code{url}, the URL of the license description.
}
}
\description{
This function returns a list of all available licenses. This is particularly
useful to get the right spelling of the license to be passed to
\code{\link[=new_package]{new_package()}}, \code{\link[=new_compendium]{new_compendium()}}, or \code{\link[=add_license]{add_license()}}.
}
\examples{
get_licenses()
}
\seealso{
Other utilities functions: 
\code{\link{get_all_dependencies}()},
\code{\link{get_all_functions}()},
\code{\link{get_minimal_r_version}()}
}
\concept{utilities functions}
