% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rd_families.R
\name{get_rd_families}
\alias{get_rd_families}
\title{List all Rd function families}
\usage{
get_rd_families()
}
\value{
A vector of Rd function families.
}
\description{
This function parses all R functions to detect \verb{@family} tag.
}
\examples{
\dontrun{
get_rd_families()
}
}
\seealso{
Other utilities functions: 
\code{\link{get_all_dependencies}()},
\code{\link{get_all_functions}()},
\code{\link{get_minimal_r_version}()}
}
\concept{utilities functions}
