% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_compendium.R
\name{new_compendium}
\alias{new_compendium}
\title{Create a Research Compendium}
\usage{
new_compendium(overwrite = FALSE)
}
\arguments{
\item{overwrite}{a logical value. If a \code{make.R} is already present and
\code{overwrite = TRUE}, this file will be erased and replaced.}
}
\description{
This function creates a new files and folders structure according to the
research compendium paradigm. The project will be installed as a package.
}
