% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompendium-package.R
\docType{package}
\name{rcompendium-package}
\alias{rcompendium}
\alias{rcompendium-package}
\title{rcompendium: Create a Research Compendium}
\description{
This package creates a research compendium, i.e. a predefined
    folder structure. It can be also used to create a new R package as an
    alternative to usethis::create_package().
}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
