% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompendium-package.R
\docType{package}
\name{rcompendium-package}
\alias{rcompendium}
\alias{rcompendium-package}
\title{rcompendium: Create a Package or Research Compendium Structure}
\description{
The purpose of the package \code{rcompendium} is to make easier
    the creation of R package/research compendium (i.e. a predefined
    files/folders structure) so that user can focus on the code/analysis
    instead of wasting time organizing files. A full ready-to-work structure is
    set up with some additional features (versioning, GitHub repository
    creation, GitHub Actions configuration to automatically check package
    integrity and deploy website using \code{pkgdown}). This package relies
    heavily on the packages \code{devtools} and \code{usethis}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/rcompendium}
  \item Report bugs at \url{https://github.com/FRBCesab/rcompendium/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
