% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refresh a package/research compendium}
\usage{
refresh(quiet = FALSE, check = TRUE, import = NULL)
}
\arguments{
\item{quiet}{a logical value. If \code{TRUE} (default) no message are printed.}

\item{check}{a logical value. If \code{TRUE} (default) package integrity is
checked using \code{\link[devtools:check]{devtools::check()}}.}

\item{import}{(optional) the name of the folder to recursively detect
dependencies to be added in the \code{Imports} field of \strong{DESCRIPTION} file.
Default is \code{import = NULL} (but \strong{R/}, \strong{NAMESPACE}, and \verb{@examples} are
still inspected).}
}
\description{
This function refreshes a package/research compendium. It will:
\itemize{
\item Update \code{.Rd} files and \code{NAMESPACE}, using \code{\link[devtools:document]{devtools::document()}};
\item Update external packages (in \code{DESCRIPTION} file);
\item Update badges in \code{README.Rmd} (if already present);
\item Re-knitr the \code{README.Rmd}, using \code{\link[rmarkdown:render]{rmarkdown::render()}};
\item Check package integrity with \code{\link[devtools:check]{devtools::check()}}.
}
}
\examples{
\dontrun{
rcompendium::refresh()
}
}
\seealso{
Other setup functions: 
\code{\link{new_compendium}()},
\code{\link{new_package}()},
\code{\link{set_credentials}()}
}
\concept{setup functions}
