% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Store user information in the .Rprofile}
\usage{
set_credentials(
  given = NULL,
  family = NULL,
  email = NULL,
  orcid = NULL,
  github_user = NULL,
  protocol = "https",
  open = TRUE
)
}
\arguments{
\item{given}{a \code{character} of length 1. The given name of the user
(considered as the maintainer and code owner of the project).}

\item{family}{a \code{character} of length 1. The family name of the user
(considered as the maintainer and code owner of the project).}

\item{email}{a \code{character} of length 1. The email address of the user
(considered as the maintainer and code owner of the project).}

\item{orcid}{a \code{character} of length 1. The ORCID of the user
(considered as the maintainer and code owner of the project).}

\item{github_user}{a \code{character} of length 1. The GitHub account name of the
user (considered as the maintainer and code owner of the project).}

\item{protocol}{a \code{character} of length 1. The GIT protocol used to
communicate with GitHub. One of \code{'https'} or \code{'ssh'}. If you
don't know, keep the default value (i.e. \code{https}).}

\item{open}{a \code{logical} value. If \code{TRUE} (default) the \code{.Rprofile} is opened
in the default text editor (recommended).}
}
\value{
No return value.
}
\description{
This function is used to store user credentials (given and family names,
email, ORCID, GitHub username, etc.) in the \code{.Rprofile} file.
This function is useful to store user credentials once for all especially if
the user creates a lot of R projects and/or if the user manually calls
the \verb{add_*()} functions.

This function will create the \code{.Rprofile} file if it does not exist. Then,
the user needs to paste the content of the clipboard to this file (open by
the function).
}
\examples{
\dontrun{
set_credentials(
  given = "John",
  family = "Doe",
  email = "john.doe@domain.com",
  orcid = "9999-9999-9999-9999",
  github_user = "jdoe",
  protocol = "https",
  open = TRUE
)
}
}
\seealso{
Other setup functions: 
\code{\link{new_compendium}()},
\code{\link{new_package}()}
}
\concept{setup functions}
