% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Store credentials to the .RProfile}
\usage{
set_credentials(
  given = NULL,
  family = NULL,
  email = NULL,
  orcid = NULL,
  protocol = NULL,
  open = FALSE
)
}
\arguments{
\item{given}{(optional) given name of the package maintainer}

\item{family}{(optional) family name of the package maintainer}

\item{email}{(optional) email address of the package maintainer}

\item{orcid}{(optional) ORCID of the package maintainer}

\item{protocol}{(optional) GIT protocol used to communicate with the GitHub
remote. One of 'https' or 'ssh'. If you don't know, let \code{protocol = NULL}
and the protocol will be 'https'.}

\item{open}{a logical value. If \code{TRUE} the file is opened in the editor.
Default is \code{open = FALSE}}
}
\description{
This function stores user credentials in the \code{.Rprofile} file. Accepted
credentials are listed below. This function is useful if user creates a lot
of packages and/or research compendiums.

If the \code{.Rprofile} file does not exist this function will create it. Note
that credentials are added at the end of the file: if you run this function
two times, credentials will be added twice. User can run the command
\code{set_credentials(open = TRUE)} to open the \code{.Rprofile} and clean its
content. \strong{Be careful while modifying this file!}.
}
\examples{
\dontrun{
library(rcompendium)


## Define **ONCE FOR ALL** your credentials ----

set_credentials("John", "Doe", "john.doe@domain.com", 
                orcid = "9999-9999-9999-9999", protocol = "ssh")
}
}
\seealso{
Other setup functions: 
\code{\link{new_compendium}()},
\code{\link{new_package}()},
\code{\link{refresh}()}
}
\concept{setup functions}
