% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Store credentials to the .RProfile}
\usage{
set_credentials(
  given = NULL,
  family = NULL,
  short = NULL,
  email = NULL,
  orcid = NULL,
  github = NULL,
  license = NULL
)
}
\arguments{
\item{given}{(optional) user given name (i.e. "Nicolas")}

\item{family}{(optional) user family name (i.e. "Casajus")}

\item{short}{(optional) user short name (e.g. "Casajus N.")}

\item{email}{(optional) user email address (i.e. "nicolas.casajus@gmail.com")}

\item{orcid}{(optional) user ORCID (i.e. "0000-0002-5537-5294")}

\item{github}{(optional) user GitHub pseudo/organization (i.e. "ahasverus")}

\item{license}{(optional) user preferred license (i.e. "GPL (>= 2)")}
}
\description{
This function adds user credentials in the \code{.Rprofile} file. Accepted
credentials are listed below. This function is useful if user creates a lot
of packages and/or research compendiums.

If the \code{.Rprofile} file does not exist this function will create it. Note
that credentials are added at the end of the file: if you run this function
two times, credentials will be added twice. User can run the command
\code{usethis::edit_r_profile()} to open the \code{.Rprofile} and clean its content.
\strong{Be careful}.
}
\examples{
\dontrun{
set_credentials(given = "Nicolas", family = "Casajus", short = "Casajus N.",
                email = "nicolas.casajus@gmail.com", github = "ahasverus", 
                orcid = "0000-0002-5537-5294", license = "GPL (>= 2)")
}
}
\seealso{
\code{\link[=new_compendium]{new_compendium()}}, \code{\link[=new_package]{new_package()}}

Other core functions: 
\code{\link{new_compendium}()}
}
\concept{core functions}
